updated index files and created test and code files
