+++ lib/frogger.js
@@ -1,0 +1,5 @@
+function Frogger(){
+
+}
+
+module.exports = Frogger;
