+++ lib/index.js
@@ -1,8 +1,2 @@
 var canvas = document.getElementById('canvas');
 var context =  canvas.getContext('2d');
-
-function Frogger(){
-  
-}
-
-module.exports = index.js
