+++ test/index.js
@@ -1,16 +1,1 @@
-var assert = require('chai').assert;
-var Frogger = require('../lib/index');
-
-describe('Frogger', function() {
-
-  it('should have a frog', function() {
-    assert.isFunction(Frogger);
-  });
-
-  it('should instantiate the first round', function(){
-    var frog = new Frogger();
-    assert.isObject(frog);
-  });
-
-  
-})
+require('./frogger-test.js');
